Add scratch directory to heron docker build
This sets the scratch directory for a build to be $HOME/.heron-compile.
The purpose of this is to not add files to the working directory that
might accidentally get committed while a build is taking place.

This change also checks to make sure the output directory exists and
creates it if it does not already exist.