Fix  UnsupportedOperationException in AuroraUtils
`Arrays.asList` returns an immutable abstract implementation of list and
cannot have elements added to it. This change wraps the array list
returned in a mutable copy of the list to allow adding more arguments to
the auroraCmd.

```
java.lang.UnsupportedOperationException
        at java.util.AbstractList.add(AbstractList.java:148)
        at java.util.AbstractList.add(AbstractList.java:108)
        at com.twitter.heron.scheduler.aurora.AuroraUtils.createAuroraJob(AuroraUtils.java:23)
        at com.twitter.heron.scheduler.aurora.AuroraLauncher.launch(AuroraLauncher.java:124)
        at com.twitter.heron.scheduler.LaunchRunner.call(LaunchRunner.java:128)
        at com.twitter.heron.scheduler.SubmitterMain.submitTopology(SubmitterMain.java:376)
        at com.twitter.heron.scheduler.SubmitterMain.main(SubmitterMain.java:306)
```

This also fixes exception when casting the list to a string array.
`List.toArray` also returns an Object[] instead of a typed array. By
adding a typed array to the `List.toArray` argument list it will convert
it to the approriate type and does not need to be cast.

```
java.lang.ClassCastException: [Ljava.lang.Object; cannot be cast to [Ljava.lang.String;
        at com.twitter.heron.scheduler.aurora.AuroraUtils.createAuroraJob(AuroraUtils.java:38)
        at com.twitter.heron.scheduler.aurora.AuroraLauncher.launch(AuroraLauncher.java:124)
        at com.twitter.heron.scheduler.LaunchRunner.call(LaunchRunner.java:128)
        at com.twitter.heron.scheduler.SubmitterMain.submitTopology(SubmitterMain.java:376)
        at com.twitter.heron.scheduler.SubmitterMain.main(SubmitterMain.java:306)
```