Set ExecutionState completely in LauncherRunner and Refactor ILauncher interface  (#305)
Currently ILauncher implementation is expected to update details in ExecutionState, which may not be so intutitive (see LocalLauncher.updateExecutionState).heron-tracker ignores any topologies available in state manager if the cluster and role info is not provided by ILaucher implementation.

LaunchRunner has complete information, including cluster and others, and it can be set it at creation time. It reduces the complexity to implement ILauncher.