Fix concurrent bugs in Scheduler
Heron check whether topology def exists in StateManager to determine whether a topology already exists.
So it is needed to write topology def to StateManager first when submitting a topology,
and delete topology def at the end when killing a topology.
Otherwise, there could be concurrent modification on StateManager node,
when there are multiple topologies submitted at the same time.