Make sure permissions are set before writing artifacts
Bazel defaults all packages to 0555 permissions which means they are
readable and executable but not writable. If you run the compile script
twice in a row or any of the artifacts already exist in the destination
folder the copy will fail because ther are no write privileges. This
change makes sure that if the file exists it has sufficient permissions
to allow for the script to overwrite it.