Upgrade httpclient and httpcore version
These two dependencies are versioned separately so I remove the
httpcomponents_version designation in favor of tracking their versions
independently of each other. The main reason for this is http_client has a
higher version number than http_core and the newer version is needed for
the aws s3 sdk.