Update aws dependency version (#1042)
* Update aws dependency version

This is to address a 400 Bad Request exception coming from the s3 client
in some cases: https://github.com/twitter/heron/issues/1035

* Upgrade httpclient and httpcore version

These two dependencies are versioned separately so I remove the
httpcomponents_version designation in favor of tracking their versions
independently of each other. The main reason for this is http_client has a
higher version number than http_core and the newer version is needed for
the aws s3 sdk.