Do not let external processes block on IO buffers in ShellUtils.runSyncProcess
In ShellUtils.runSyncProcess, "output from the external process can exhaust the available buffer for the output or error stream because neither stream is read while waiting for the process to complete. If either buffer becomes full, it can block the external process as well, preventing all progress for both the external process and the Java program." [1]

This PR just launches threads to read outputs asynchronously to fix this issue. The added unit tests can reproduce this bug.

[1] https://www.securecoding.cert.org/confluence/display/java/FIO07-J.+Do+not+let+external+processes+block+on+IO+buffers