Wait for a new pplan properly (#1605)
In test case StMgr.test_tmaster_restart_on_same_address, currently we wait for 3 seconds for a new pplan to come in and then check for the new pplan. This is broken if we don't receive the new pplan within 3 seconds. So, just wait for the change of the pplan and timeout after 30 seconds.