Stream subprocess stdout and stderr in heron-executor
Previously the logs for subprocesses launched by the heron-executor were
not actually streamed to the heron-executor log file. Using
`communicate` to read the output of a subprocess will wait until the
process is closed before actually reading the logs. In the non-blocking
case the stdout and stderr were sent to a pipe but were not actually
read from which results in a deadlock if the subprocess logs enough
information into the stderr buffer
([python docs on subprocess and
pipes](https://docs.python.org/2/library/subprocess.html#subprocess.check_call)).

> Do not use stdout=PIPE or stderr=PIPE with this function as that can
deadlock based on the child process output volume.

Heron uses glog to log messages to a file however zookeeper defaults to
using stderr with no easy way to tell it to use glog instead. As such
whenever zookeeper logged a message it would get written to the
subprocess stderr buffer and fill it up ultimately causing a deadlock in
the zookeeper thread in the tmaster trying to write to stderr. This then
manifests itself by the zookeeper client not being able to heartbeat and
having its session expire without being able to callback and restart the
process.

With `subprocess.communicate` the data is [buffered in
memory](https://docs.python.org/2/library/subprocess.html#subprocess.Popen.communicate) until end-of-line is reached
and the subprocess is terminated so it is not suitable for long running processes.

One design consideration here for simplicity is the redirection of the
subprocess stderr to stdout. This is to make reading the data easier.
The downside here is that it is not possible to tell if a log message
came from stdout or stderr but I think that information will be clear
with the content of the message. Also some dependencies like zookeeper
log all of their messages on stderr anyway regardless of the message
type.