Set the logging thread as a daemon thread
If the thread is not set as daemon it will cause the heron executor to
hang on exit waiting for the threads to finish. They will never finish
because they are waiting on the subprocess to be terminated which
happens only when the heron-executor exists. By setting them as daemon
threads python will exit the parent thread when there are only daemon
threads left and will abruptly kill them.