Retry when failing to kill an aurora topology
A lot of customers complained heron failing to kill an aurora topology because
the aurora service can be unavailable or unstable for a while, or it can provide false negative resposne.

This pull request adds retry mechanism to kill an auorora topology. People can also specify the retry attempts and interval via config.