Retry when failing to kill an Aurora topology (#2554)
A lot of Twitter internal topology owners complained heron failing to kill an aurora topology because
the Aurora service can be unavailable or unstable for a while, or it can sometimes provide a false negative response.

This pull request adds retry mechanism to kill an Aurora topology with more confidence. It can increase the robustness of AuroraScheduler regardless of the Aurora Service People can also specify the retry attempts and interval via config.

This pull request also re-format some coding-styles for RuntimeManagerRunner.java.

Tested with actual Twitter distributed env.