Log component's config when constructing the instance
Currently, people are having a hard time to figure out the actual configs applying to a component:
- There are a lot of transformations and overrides
- The actual configs applied are not directly visible

This pull request logs the actual configs applying to the instance during the actual config construction for an instance.
It makes debug easiler.
Furthermore, it is invoked only when having new PhysicalPlanHelper, so it does not hurt runtime performance.