Add a --heron-home option in addition to --config-path option in heron-cli.
This new option can be used to override libraries and configs stored in
heron-cli.

The usage is pretty much the same as the existing --config-path option.
For example, if a user needs to manupulate an old topology, which has
imcompatible libraries. User can download an old version of heron-cli
and extract libraries then use this option to specify the path to it.