Clean up simulator/utils
This is the first commit i'll be making to the simulator package, as I attempt to add stateful processing to the simulator pursuant to issue #2346.

This cleanup first removes a couple of unused classes from simulator/instance.

It also removes some odd programming paradigms which were used for the simulator utils, like a pseudo-static class that acts on an object and objects derived from it (PhysicalPlanUtil), a class that can only be instantiated by a static method which creates maps of itself (StreamConsumers), a class with an entirely unnecessary private constructor (XORManager) and general passing of directly derived objects instead of encapsulation.  I've merged the PhysicalPlanUtil and StreamConsumers functionality into TopologyManager, a class which manages additional functionality for a single topology and its directly derived objects.