Add --skip-runtime-validation option in heron kill command
Currently kill command tries to kill cluster and then clear state
data. In a relatively rare case, the clear state action might not be fully
successful, and as the result, user might want to kill the topology
again in order to clear the data left. But runtime manager object tries
to validate the topology is available or not first and stops the
process.