Add sha256 checksums for all http_archives (#2856)
I have taken the current sha256 checksum for all of the workspace
archives so we can verify their signature on download. Since some
of the artifacts are downloaded over http, we want to make sure
there was no man in the middle attack done to change the resulting
binary. This also  gives assurance that the code we are downloading
has not been tampered with in any way either over the wire or at
the source.