S3Uploader bug fixes
This fixes 2 bugs in the S3Uploader.

The first is when setting the region it is not checked before-hand if
the region is set or not. In the event that the end user does not
explicitly set the region in the config it will automatically be set to
`null`. Since `null` is not a valid s3 region the uploader will fail
with a 301 on any command which can be confusing. It is not required
that the s3 client sets a region and in the event that one is not set
explicitly it will default to us-east-1.

The second is the `close` method does not check to see if the file
exists before it deletes it. In the event that there is a failure
earlier in the uploading process the `previousVersionFilePath` will
exist but was never uploaded to s3 and the `close` method will throw
an exception.