Update WORKSPACE info for zookeeper http archive. (#3205)
* Update WORKSPACE info for zookeeper http archive.

Updated the new_http_archive entry for org_apache_zookeeper to
zookeeper-3.4.13. Performing a clean installation of Heron failed due to
the fact that the previous version (3.4.10) is no longer accessible at
the url locations provided.

* Update WORKSPACE info for zookeeper http archive.

Modified to use archive.apache.org as the url for zookeeper.
Reverted to original version of 3.4.10. Using archive.apache.org should
be sufficient for retrieving zookeeper versions in the future without
fear of the previous mirror sites removing versions needed by the
WORKSPACE file.