[STREAMCOMP-2885] broadcast GCC completition (#3330)
* amend gitignore

* decouple GCC post-save logic from stateful controller

* tmaster broadcast checkpoint completion

the message will be first sent to all stmgrs, and then each stmgr will forward
it to every heron instance it connects to.

* heron instance handles global checkpoint saved msg

* expose necessary API for the 2PC support

unit tests for this will be in the following commit

* test: refactor mock physical plan to a builder

* test preSave and postSave hook

* [STREAMCOMP-2916] block execute on postSave

* add stateful tests for spouts

* test preRestore

* rename interface to ITwoPhaseStatefulComponent

* block spout from reading data tuples as well

* fix log format

* do not check interface

this is because the boolean will only be set to true for tasks that implement ITwoPhaseStatefulComponent. No need to do the check again.