Remove stray space characters between \ and \n (#3843)
The example bash shell commands in this document are wrapped onto multiple lines with
backslashes for legibility. However, there are space characters inserted after the backslash
escapes, breaking the commands: the spaces are being escaped, not the newlines. An
attempt to copy and paste the commands directly from the documentation will fail.

Remove the spaces to fix the commands.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>